requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1
httpx==0.27.0
pydantic==2.6.4
psycopg2-binary==2.9.9